requirements.txt
APScheduler==3.6.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
configparser==3.7.4
Django==2.2.2
django-rest==0.0.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
freeze==1.0.10
html5lib==1.0.1
idna==2.8
kronos==0.2.2
lxml==4.3.3
MechanicalSoup==0.11.0
mechanize==0.4.2
mysqlclient==1.4.2.post1
pycrypto==2.6.1
pytz==2019.1
requests==2.21.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
ssh==1.8.0
tzlocal==1.5.1
urllib3==1.24.3
webencodings==0.5.1
